# ext/random: Optimized `getBytes` loop processing