--- a/ext/random/php_random.h
+++ b/ext/random/php_random.h
@@ -217,30 +217,16 @@ PHPAPI ZEND_EXTERN_MODULE_GLOBALS(random)
 /* Bytes swap */
 #ifdef _MSC_VER
 #  include <stdlib.h>
-#  define RANDOM_BSWAP32(u) _byteswap_ulong(u)
 #  define RANDOM_BSWAP64(u) _byteswap_uint64(u)
 #else
 #  ifdef __GNUC__
-#    define RANDOM_BSWAP32(u) __builtin_bswap32(u)
 #    define RANDOM_BSWAP64(u) __builtin_bswap64(u)
 #  elif defined(__has_builtin)
-#    if __has_builtin(__builtin_bswap32)
-#      define RANDOM_BSWAP32(u) __builtin_bswap32(u)
-#    endif // __has_builtin(__builtin_bswap32)
 #    if __has_builtin(__builtin_bswap64)
 #      define RANDOM_BSWAP64(u) __builtin_bswap64(u)
 #    endif // __has_builtin(__builtin_bswap64)
 #  endif // __GNUC__
 #endif // _MSC_VER
-#ifndef RANDOM_BSWAP32
-static inline uint32_t RANDOM_BSWAP32(uint32_t u)
-{
-  return (((u & 0xff000000) >> 24)
-          | ((u & 0x00ff0000) >>  8)
-          | ((u & 0x0000ff00) <<  8)
-          | ((u & 0x000000ff) << 24));
-}
-#endif
 #ifndef RANDOM_BSWAP64
 static inline uint64_t RANDOM_BSWAP64(uint64_t u)
 {