--- a/ext/random/randomizer.c
+++ b/ext/random/randomizer.c
@@ -278,6 +278,7 @@ PHP_METHOD(Random_Randomizer, getBytes)
 
 	zend_string *retval;
 	zend_long user_length;
+	size_t total_size = 0;
 
 	ZEND_PARSE_PARAMETERS_START(1, 1)
 		Z_PARAM_LONG(user_length)
@@ -290,32 +291,17 @@ PHP_METHOD(Random_Randomizer, getBytes)
 
 	size_t length = (size_t)user_length;
 	retval = zend_string_alloc(length, 0);
-	char *rptr = ZSTR_VAL(retval);
 
-	size_t to_read = length;
-	while (to_read > 0) {
+	while (total_size < length) {
 		php_random_result result = engine.algo->generate(engine.state);
 		if (EG(exception)) {
 			zend_string_free(retval);
 			RETURN_THROWS();
 		}
-
-		uint64_t tmp_ret = result.result;
-		if (to_read >= result.size && result.size == sizeof(uint64_t)) {
-#ifdef WORDS_BIGENDIAN
-			tmp_ret = RANDOM_BSWAP64(tmp_ret);
-#endif
-			memcpy(rptr, &tmp_ret, sizeof(uint64_t));
-			to_read -= sizeof(uint64_t);
-			rptr += sizeof(uint64_t);
-		} else {
-			for (size_t i = 0; i < result.size; i++) {
-				*rptr++ = tmp_ret & 0xff;
-				tmp_ret >>= 8;
-				to_read--;
-				if (to_read == 0) {
-					break;
-				}
+		for (size_t i = 0; i < result.size; i++) {
+			ZSTR_VAL(retval)[total_size++] = (result.result >> (i * 8)) & 0xff;
+			if (total_size >= length) {
+				break;
 			}
 		}
 	}