@SakiTakamachi 2024-07-09
@TimWolla 
Thanks for your review.

> I'm not too happy about the additional complexity introduced here and I would prefer a solution where the compiler itself detects that this effectively is a memcpy() for little-endian architectures, while the C code stays high level and readable.

I could probably simplify the code a bit more, but it would still be somewhat complicated.

I'm going to simplify it a bit more so you can check back later to see if it's satisfactory to you :)