@SakiTakamachi 2024-07-09
@TimWolla 
I've simplified it as much as possible.