@SakiTakamachi 2024-07-11
Of course, if there is a more clever solution I'll be happy to close this PR.