@SakiTakamachi 2024-07-11
Split the changes into multiple commits to make it easier to see the changes.
The measurement results for each change will be described in the explanation.
(The code has changed slightly.)