@TimWolla 2024-08-04
I have filed a competing PR at #15228.